chore: moved useViewer types and utils to their own files
